Remedial measure: Readd highlight colors to the plugin
I wanted to have highlight colors that would fit in with my
colorscheme, but discovered that I couldn't override in my vimrc the
highlights that were already set in the plugin file. Until I find out
how to do this, I will add these colors in the plugin file, replacing the previous setting.